+++ src/components/Shared/StatusFilter/StatusFilter.module.css
@@ -1,0 +1,19 @@
+.option {
+  padding: 6px 10px;
+}
+
+.container {
+  display: flex;
+  align-items: center;
+  gap: 8px;
+}
+
+.icon {
+  width: 16px;
+  height: 16px;
+  flex-shrink: 0;
+}
+
+.label {
+  font-size: 14px;
+}
